refactor: move respective components under Main component
