+++ components/Main/Rule/Rule.module.css
