+++ components/Main/Rule/index.tsx
